requirements.txt
# Flask Web框架
Flask==2.3.3
flask-cors==4.0.0

# 环境变量管理
python-dotenv==1.0.0

# HTTP请求
requests==2.31.0

# 工具库
werkzeug==2.3.7